Medal/finance-copilot/requirements.txt
flask>=3.0,<4
pandas>=2.2,<3
openai>=1.40,<2
google-generativeai>=0.8,<1
gunicorn>=22,<24
